+++ pkg/providers/property/resource_akamai_property_include_activation.go
@@ -101,10 +101,31 @@ func resourcePropertyIncludeActivation() *schema.Resource {
 				Description: "Provides an audit record when activating on a production network",
 				Elem:        complianceRecordSchema,
 			},
+			"timeouts": {
+				Type:        schema.TypeList,
+				Optional:    true,
+				MaxItems:    1,
+				Description: "Enables to set timeout for processing",
+				Elem: &schema.Resource{
+					Schema: map[string]*schema.Schema{
+						"default": {
+							Type:             schema.TypeString,
+							Optional:         true,
+							ValidateDiagFunc: validateDurationFormat,
+						},
+					},
+				},
+			},
 		},
 		Timeouts: &schema.ResourceTimeout{
 			Default: readTimeoutFromEnvOrDefault("AKAMAI_ACTIVATION_TIMEOUT", includeActivationTimeout),
 		},
+		SchemaVersion: 1,
+		StateUpgraders: []schema.StateUpgrader{{
+			Version: 0,
+			Type:    resourcePropertyIncludeActivationV0().CoreConfigSchema().ImpliedType(),
+			Upgrade: migrateTimeoutsToCustom(),
+		}},
 	}
 }
 
@@ -209,6 +230,11 @@ func resourcePropertyIncludeActivationUpdate(ctx context.Context, d *schema.Reso
 	if mutableAttrsHaveChanges && !d.HasChanges("version") {
 		return diag.FromErr(fmt.Errorf("attributes such as 'notify_emails', 'auto_acknowledge_rule_warnings', " +
 			"'compliance_record' cannot be updated after resource creation without 'version' attribute modification"))
+	}
+
+	if !d.HasChangeExcept("timeouts") {
+		logger.Debug("Only timeouts were updated, skipping")
+		return nil
 	}
 
 	err := resourcePropertyIncludeActivationUpsert(ctx, d, client)